Moving views bindings module into new place.
