+++ graylog-storage-elasticsearch6/src/main/java/org/graylog/storage/elasticsearch6/Elasticsearch6Plugin.java
@@ -1,11 +1,11 @@
 package org.graylog.storage.elasticsearch6;
 
+import com.google.common.collect.ImmutableList;
 import org.graylog2.plugin.Plugin;
 import org.graylog2.plugin.PluginMetaData;
 import org.graylog2.plugin.PluginModule;
 
 import java.util.Collection;
-import java.util.Collections;
 
 public class Elasticsearch6Plugin implements Plugin {
     @Override
@@ -15,6 +15,9 @@ public class Elasticsearch6Plugin implements Plugin {
 
     @Override
     public Collection<PluginModule> modules() {
-        return Collections.singleton(new Elasticsearch6Module());
+        return ImmutableList.of(
+                new Elasticsearch6Module(),
+                new ViewsESBackendModule()
+        );
     }
 }
